learntools/airflow/learning-airflow/requirements.txt
# Add Python packages here, one per line. Pin versions where possible.
snowflake-connector-python==3.6.0
apache-airflow-providers-snowflake==5.1.0

